+++ src/services/__mocks__/order/schema.js
@@ -4,11 +4,22 @@ export const orderSchema = ({ ...props }) =>
     id: props.id,
     display_id: props.display_id,
     cart_id: props.cart_id,
-    shipping_address: {},
+    shipping_address: props.shipping_address ?? {},
     fulfillments: props.fulfillments,
     claims: [],
     swaps: [],
     items: props.items,
+    region: Object.freeze({
+      currency_code: "usd",
+    }),
+    shipping_methods: [
+      Object.freeze({
+        shipping_option: Object.freeze({
+          name: "Just best, more better",
+        }),
+      }),
+    ],
+    currency_code: "usd",
     shipping_total: 1793,
     tax_total: 899,
     subtotal: 7200,