+++ src/services/__tests__/setup/index.js
@@ -1,5 +1,6 @@
 export {
   makeShippoClientService,
+  makeShippoFulfillmentService,
   makeShippoPackerService,
   makeShippoPackageService,
   makeShippoRatesService,